feat: add @ToString for models
