+++ build.gradle
@@ -12,7 +12,6 @@ repositories {
 
 dependencies {
     implementation 'org.apache.commons:commons-lang3:3.14.0'
-    implementation 'org.apache.commons:commons-collections4:4.4'
     implementation 'org.apache.logging.log4j:log4j-core:2.23.1'
     implementation 'com.github.javafaker:javafaker:1.0.2'
     implementation 'com.fasterxml.jackson.core:jackson-databind:2.17.0'
