--- src/main/java/org/example/animal/pet/Cat.java
@@ -1,5 +1,0 @@
-package org.example.animal.pet;
-
-public class Cat extends Pet {
-
-}
