--- src/main/java/org/example/animal/pet/Dog.java
@@ -1,5 +1,0 @@
-package org.example.animal.pet;
-
-public class Dog extends Pet {
-
-}
