--- src/main/java/org/example/animal/predator/Shark.java
@@ -1,5 +1,0 @@
-package org.example.animal.predator;
-
-public class Shark extends Predator {
-
-}
