--- src/main/java/org/example/animal/predator/Wolf.java
@@ -1,5 +1,0 @@
-package org.example.animal.predator;
-
-public class Wolf extends Predator {
-
-}
