+++ src/main/java/org/example/animals/pets/Cat.java
@@ -1,0 +1,8 @@
+package org.example.animals.pets;
+
+import lombok.ToString;
+
+@ToString(callSuper = true)
+public class Cat extends Pet {
+
+}
