+++ src/main/java/org/example/animals/predators/Shark.java
@@ -1,0 +1,8 @@
+package org.example.animals.predators;
+
+import lombok.ToString;
+
+@ToString(callSuper = true)
+public class Shark extends Predator {
+
+}
