+++ src/main/java/org/example/animals/predators/Wolf.java
@@ -1,0 +1,8 @@
+package org.example.animals.predators;
+
+import lombok.ToString;
+
+@ToString(callSuper = true)
+public class Wolf extends Predator {
+
+}
