+++ src/main/java/org/example/services/impl/SearchServiceImpl.java
@@ -1,9 +1,9 @@
-package org.example.service.impl;
+package org.example.services.impl;
 
-import org.example.animal.AbstractAnimal;
+import org.example.animals.AbstractAnimal;
 import org.example.errors.InvalidAnimalBirthDateException;
 import org.example.errors.InvalidAnimalException;
-import org.example.service.SearchService;
+import org.example.services.SearchService;
 
 public class SearchServiceImpl implements SearchService {
