+++ src/test/java/org/example/animals/repositories/AnimalRepositoryImplTest.java
@@ -1,11 +1,10 @@
-package org.example.animal.repository;
+package org.example.animals.repositories;
 
-import org.example.animal.AbstractAnimal;
-import org.example.animal.pet.Cat;
-import org.example.animal.pet.Dog;
-import org.example.animal.predator.Shark;
-import org.example.animal.predator.Wolf;
-import org.example.utils.ResultReader;
+import org.example.animals.AbstractAnimal;
+import org.example.animals.pets.Cat;
+import org.example.animals.pets.Dog;
+import org.example.animals.predators.Shark;
+import org.example.animals.predators.Wolf;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -13,13 +12,11 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 
 import java.io.ByteArrayOutputStream;
-import java.io.FileNotFoundException;
 import java.io.PrintStream;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -86,37 +83,19 @@ class AnimalRepositoryImplTest {
 
     @Test
     @DisplayName("Позитивный тест findOlderAnimal с возрастом меньше возраста одного из животных")
-    void successFindOlderAnimals() throws FileNotFoundException {
+    void successFindOlderAnimal() {
         int age = LocalDate.now().getYear() - wolf.getBirthDate().getYear();
         Map<AbstractAnimal, Integer> olderAnimals = animalRepository.findOlderAnimal(animals, age);
         assertThat(olderAnimals).hasSize(4)
                 .doesNotContainKey(wolf);
-        checkFindOlderAnimalsJson(olderAnimals);
-    }
-
-    private void checkFindOlderAnimalsJson(Map<AbstractAnimal, Integer> olderAnimals) throws FileNotFoundException {
-        List<AbstractAnimal> abstractAnimals = ResultReader.readOlderAnimals();
-        assertThat(olderAnimals).hasSameSizeAs(abstractAnimals);
-
-        for (AbstractAnimal abstractAnimal : abstractAnimals) {
-            boolean isMatchedAnimal = olderAnimals.keySet().stream()
-                    .filter(animal -> animal.getName().equals(abstractAnimal.getName()))
-                    .allMatch(animal -> Objects.equals(animal.getBreed(), abstractAnimal.getBreed()) &&
-                            Objects.equals(animal.getCost(), abstractAnimal.getCost()) &&
-                            Objects.equals(animal.getCharacter(), abstractAnimal.getCharacter()) &&
-                            Objects.equals(animal.getBirthDate(), abstractAnimal.getBirthDate()) &&
-                            Objects.equals(animal.getSecretInformation(), abstractAnimal.getSecretInformation()));
-            assertThat(isMatchedAnimal).isTrue();
-        }
     }
 
     @Test
     @DisplayName("Позитивный тест findOlderAnimal с возрастом больше возраста любого животного")
-    void successFindOlderAnimalOfPossible() throws FileNotFoundException {
+    void successFindOlderAnimalOfPossible() {
         Map<AbstractAnimal, Integer> olderAnimals = animalRepository.findOlderAnimal(animals, Integer.MAX_VALUE);
         int expectedAge = LocalDate.now().getYear() - cat1.getBirthDate().getYear();
         assertThat(olderAnimals).containsExactlyEntriesOf(Map.of(cat1, expectedAge));
-        checkFindOlderAnimalsJson(olderAnimals);
     }
 
     @ParameterizedTest(name = "{displayName} [{index}] animals={0}")
